requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
billiard==3.6.4.0
boto3==1.36.6
CacheControl==0.14.0
cachetools==5.3.3
celery==5.2.7
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.38.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
dj-database-url==1.3.0
Django==4.2.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.0
django-environ==0.10.0
django-storages==1.14.4
djangorestframework==3.14.0
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
fcm-django==2.0.1
firebase-admin==6.5.0
grpcio==1.62.2
grpcio-status==1.62.2
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.2.4
MarkupSafe==2.1.3
msgpack==1.0.8
multidict==6.0.4
openai==0.27.2
packaging==23.1
paypack-py==1.0.4
Pillow==10.0.0
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.10.0
pycparser==2.22
pyenchant==3.2.2
PyJWT==2.9.0
pyparsing==3.1.2
pytz==2022.7.1
PyYAML==6.0.1
redis==4.5.4
referencing==0.35.1
requests==2.26.0
rpds-py==0.20.0
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.48
sqlalchemy-redshift==0.8.14
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
